requirements.txt
textual>=0.85.0
httpx>=0.27.0
python-dotenv>=1.0.0
